+++ docs/HISTORY.md
@@ -15,10 +15,13 @@
 - Bumped versions in CI pipeline
     - uv version
     - Docker build/push version
-- Improved error messages for failed connections to Home Assistant
+- Improved error messages
+    - for failed connections to Home Assistant
+    - for failed HTTP requests to Home Assistant
 - Improved error messages for custom plugins
 - Parsing various timedeltas in config with `utils.parse_timedelta`
 - Add callback argument to Dashboard's call_service - contributed by [psolyca](https://github.com/psolyca)
+- Added docstrings to `HassPlugin` methods and added it to the reference in the docs.
 
 **Fixes**